Moved the sidebar width logic
